Add files via upload
Minor changes to the pre- and post-scrabble tasks and visual foraging task. First, the submission time of each (in)correct word is now registered in both pre- and post-scrabble tasks. Second, the direction of the polygon (self.agent.direction in visual_foraging.py) is saved in each visual_foraging_path output file. Both changes were made based on the additional analyses that were reported by Hills, Todd and Goldstone (2010). 

Finally, which is more an aesthetic component, I removed the "Wacht op de volgende letterset" text which was displayed after pressing the "Volgende letterset" button and I replaced it with an empty string (""). Change was made based the original experiment by Hills and colleagues (2010).